server/requirements.txt
alembic==1.4.0
arrow==0.15.7
asn1crypto==0.24.0
bcrypt==3.1.4
blinker==1.4
boto==2.49.0
boto3==1.12.5
botocore==1.15.5
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.2
cryptography==2.4.2
cssmin==0.2.0
dnspython==1.16.0
docutils==0.15.2
dynamodb-json==1.3
elasticsearch==7.6.0
fabric==2.4.0
Flask==1.0.2
Flask-Cors==3.0.7
flask-dynamo==0.1.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-MySQL==1.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
fsspec==0.7.3
idna==2.8
invoke==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
jsmin==2.2.2
Mako==1.1.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.19
mysqlclient==1.4.6
paramiko==2.4.2
pdf2image==1.13.1
pdfkit==0.6.1
Pillow==7.1.2
protobuf==3.6.1
pyasn1==0.4.4
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.1
python-editor==1.0.4
redis==3.5.3
Redis-Sentinel-Url==1.0.1
requests==2.22.0
rq==1.4.2
rq-dashboard==0.6.1
s3fs==0.4.2
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
SQLAlchemy==1.3.13
tqdm==4.43.0
ultimate-sitemap-parser==0.5
urllib3==1.25.8
Wand==0.6.1
webassets==0.12.1
Werkzeug==1.0.0